Port advisory map-mission scoring cluster (0x4e8750/0x4e8c20/0x4e8c50/0x4e9060/0x4e9a50) and fix five mis-conventioned callees
Re-ports the case-16 advisory scoring chain from ground truth:

- TGreatPower::ComputeAdvisoryMapNodeScoreFactorByCaseMetric (0x4e8750,
  30.91% -> 89.27%): all 7 metric cases from the jump table; arg3 is a
  TZone* (flows into metric 4's navy-order zone match and metric 7's zone
  value average), not an int nation. Old port had invented null checks,
  a wrong matrix index order in case 5, and metric-3/4 callees dispatched
  as free functions.
- TGreatPower::ComputeAdvisoryMapNodeCompositeScoreByMode (0x4e8c50,
  new port, 823B stub -> 87.44%) plus the previously unmapped 2-arg
  wrapper ComputeAdvisoryMapNodeCompositeScore (0x4e8c20, new symbols.csv
  row, 100%).
- TGreatPower::ComputeMapActionContextCompositeScoreForNation (0x4e9060,
  -> 50.33%): param retyped int -> TZone*; retired the CreateObject
  factory call for a real new TSortedByRelationshipList() +
  InitializeRelationshipRecordSize() (0x4ee5c0, new 100% port); removed
  invented guards; kept the original's dead exhausted-scan fallback.
  Residual is a register/slot-allocation cascade (structure verified).
- TAutoGreatPower::SelectAndQueueAdvisoryMapMissionsCase16 (0x4e9a50,
  fake approximation -> 85.76%): full three-tier region scan, port-zone
  context walk, defense-minister threshold table gate, war fallback that
  queues defend missions on flagged contexts. Slot-0xA2 virtual renamed
  from NoOpSlotA2 on both classes. New global
  g_afAdvisoryMissionTierThresholdByMinisterSkill_00653F18[5][6] and
  float twin g_Compute_Advisory_MinusSixFloat_00653FF8.

Calling-convention fixes proven by callee RET + callsite ECX loads
(heuristics note 112): SumNavyOrderPriorityForNation[AndNodeType] are
real TGreatPower methods taking TZone* (moved from TShip.cpp);
ComputeWeightedNeighborLinkScoreForNode is a TCountry method;
ComputeWeightedNeighborLinkScoreForNodeIndex is a TArmyMgr method (moved
from TCountry.cpp); FindMapActionContextContainingNodeByIndex and
ComputeGlobalMapActionContextNodeValueAverage are TOcean methods (moved
from TZone.cpp/free decls); GetCityIndexFromCityStatePointer is a real
1-arg __fastcall; TDiplomacyMgr::BuildRelationshipListSlot88 takes
shorts (callers push partial registers).

0x55f140 was a fabricated TGreatPower body; re-ported as the real
TZone::ComputeMapActionContextNodeValueAverage (port-zone home-region
score vs secondaryNeighbors average) using a new growth-on-read
stretch<T>::operator[] (Mac oracle name) in stretch.h. TOcean's
0x564530 now dispatches zone->... instead of the g_apNationStates
contortion, and the three mission Call30 bodies drop their
reinterpret_cast<TGreatPower*>(TZone*) casts. Removed the fake
GetMultiplierSlot21C() helper (real virtual slot 0x87 is
CountMapActionContextNodesWithNationBit).

Verified: just build green, just gates green (stub count ratcheted
1265 -> 1254), just test 109/109, stats +0.08pp average with 37
improved / 17 small drops in untouched functions (TU-growth wobble,
baseline updated). Heuristics notes 112-115 added.

+++ src/game/TDiplomacyMgr.cpp
@@ -947,14 +947,14 @@ char TDiplomacyMgr::IsPrimaryNationSlotIndex(int nationSlot) {
 }
 
 // FUNCTION: IMPERIALISM 0x004f1f70
-void TDiplomacyMgr::BuildRelationshipListSlot88(int sourceNationSlot, int primaryOnlyFlag,
+void TDiplomacyMgr::BuildRelationshipListSlot88(short sourceNationSlot, short primaryOnlyFlag,
                                                 void* listHandle) {
   // The slot signature is the native void* handle; recover the common list base once
   // (InsertCopiedRecordSortedByComparator is a TIndexAndRankList virtual, shared by every sorted-list leaf).
   TIndexAndRankList* list = static_cast<TIndexAndRankList*>(listHandle);
   short candidateNationSlot;
   short lastNationSlot;
-  if (static_cast<short>(primaryOnlyFlag) == 0) {
+  if (primaryOnlyFlag == 0) {
     candidateNationSlot = 7;
     lastNationSlot = 0x16;
   } else {
@@ -971,10 +971,10 @@ void TDiplomacyMgr::BuildRelationshipListSlot88(int sourceNationSlot, int primar
   do {
     TCountry* terrain = *terrainCursor;
     if (terrain != 0 && terrain->encodedNationSlot == -1 &&
-        candidateNationSlot != static_cast<short>(sourceNationSlot)) {
+        candidateNationSlot != sourceNationSlot) {
       RelationshipRankEntry entry;
       entry.nationSlot = candidateNationSlot;
-      int source = static_cast<short>(sourceNationSlot);
+      int source = sourceNationSlot;
       entry.standingScore =
           relationStandingScoreMatrix79c[source * kNationSlotCount + candidateIndex];
       list->InsertCopiedRecordSortedByComparator(&entry);
